Getting things out of the test file
