+++ tubul/tubul_defs.h
@@ -14,6 +14,7 @@
 #pragma once
 
 #include "tubul_types.h"
+#include "tubul_varint.h"
 #include "tubul_irange.h"
 #include "tubul_argparse.h"
 #include "tubul_time.h"
@@ -23,3 +24,5 @@
 #include "tubul_logger.h"
 #include "tubul_engine.h"
 #include "tubul_thread_pool.h"
+#include "tubul_graph.h"
+
